pyproject.toml
[project]
name = "evalset-testing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "phoenix>=0.9.1",
    "promptql-api-sdk>=0.1.1",
    "python-dotenv",
    "tabulate",
]
